moved copytree and clear_persistence to PersistenceUtil.
