+++ source/root/usr/lib/logdata-anomaly-miner/aminer/util/StringUtil.py
@@ -12,6 +12,8 @@ You should have received a copy of the GNU General Public License along with
 this program. If not, see <http://www.gnu.org/licenses/>.
 """
 import logging
+import sys
+import os
 from aminer import AMinerConfig
